+++ crates/halo/src/lib.rs
@@ -21,6 +21,9 @@
 
 extern crate alloc;
 
+#[cfg(feature = "std")]
+extern crate std;
+
 pub mod collections;
 
 #[doc(inline)]